+++ port/common/omrosbacktrace.c
@@ -44,7 +44,7 @@ struct threadData {
  * anything in the handler.
  */
 static uintptr_t
-handler(struct OMRPortLibrary *portLibrary, uint32_t gpType, void *gpInfo, void *userData)
+handler(struct OMRPortLibrary *portLibrary, uint64_t gpType, void *gpInfo, void *userData)
 {
 	return OMRPORT_SIG_EXCEPTION_RETURN;
 }
